Cleanup and fixed progress bar. Need to fix file downloading and test
