+++ lua/autorun/client/advdupe2_cl_init.lua
@@ -11,8 +11,6 @@ if(!file.Exists(AdvDupe2.DataFolder, "DATA"))then
 	file.CreateDir(AdvDupe2.DataFolder)
 end
 
-include "advdupe2/cl_file.lua"
-include "advdupe2/cl_networking.lua"
 include "advdupe2/file_browser.lua"
 include "advdupe2/sh_codec.lua"
 include "advdupe2/sh_netstream.lua"
