+++ lua/autorun/server/advdupe2_sv_init.lua
@@ -10,13 +10,9 @@ AdvDupe2.FileRenameTryLimit = 256
 include "advdupe2/sv_clipboard.lua"
 include "advdupe2/sh_codec.lua"
 include "advdupe2/sh_netstream.lua"
-include "advdupe2/sv_file.lua"
-include "advdupe2/sv_networking.lua"
 include "advdupe2/sv_misc.lua"
 
 AddCSLuaFile "autorun/client/advdupe2_cl_init.lua"
-AddCSLuaFile "advdupe2/cl_networking.lua"
-AddCSLuaFile "advdupe2/cl_file.lua"
 AddCSLuaFile "advdupe2/file_browser.lua"
 AddCSLuaFile "advdupe2/sh_codec.lua"
 AddCSLuaFile "advdupe2/sh_netstream.lua"
